Add -ensures-intervals flag
The flag limits the number of intervals being analyzed.  The limit is enforced
after every operation by a merge.

The merge targets the narrowest intervals first and merges them to an adjacent
interval with the smaller error.  If the interval to be merged is a "peak"
(both adjacent intervals have less error) then it is not merged and the next
narrowest interval is inspected.

This merge algorithm relies on non-overlapping intervals to ensure the entire
step function is not composed of peaks.  Included in this commit is a bug fix
for intervals that were overlapping on a single floating-point value.